Applied AMD64 fixes

+++ source/unix/global_os.h
@@ -1,6 +1,9 @@
-//     $Id: global_os.h,v 1.8 2004-12-16 10:22:22 mbickel Exp $
+//     $Id: global_os.h,v 1.9 2005-04-01 18:59:19 mbickel Exp $
 //
 //     $Log: not supported by cvs2svn $
+//     Revision 1.8  2004/12/16 10:22:22  mbickel
+//      Fixed compilation problems on AMD64 platform
+//
 //     Revision 1.7  2003/06/01 15:03:16  mbickel
 //      Some updates to the build system for FreeBSD
 //
@@ -64,7 +67,7 @@
   
   #define StaticClassVariable
 
-  typedef unsigned int PointerSizedInt;
+  typedef unsigned long PointerSizedInt;
 
 
   #ifdef __cplusplus
